# Improve NullHtmlEncoder perf