--- a/src/Razor/Razor/ref/Microsoft.AspNetCore.Razor.netcoreapp3.0.cs
+++ b/src/Razor/Razor/ref/Microsoft.AspNetCore.Razor.netcoreapp3.0.cs
@@ -61,17 +61,17 @@ namespace Microsoft.AspNetCore.Razor.TagHelpers
         void Init(Microsoft.AspNetCore.Razor.TagHelpers.TagHelperContext context);
         System.Threading.Tasks.Task ProcessAsync(Microsoft.AspNetCore.Razor.TagHelpers.TagHelperContext context, Microsoft.AspNetCore.Razor.TagHelpers.TagHelperOutput output);
     }
-    public partial class NullHtmlEncoder : System.Text.Encodings.Web.HtmlEncoder
+    public sealed partial class NullHtmlEncoder : System.Text.Encodings.Web.HtmlEncoder
     {
-        protected NullHtmlEncoder() { }
+        internal NullHtmlEncoder() { }
         public static new Microsoft.AspNetCore.Razor.TagHelpers.NullHtmlEncoder Default { [System.Runtime.CompilerServices.CompilerGeneratedAttribute]get { throw null; } }
         public override int MaxOutputCharactersPerInputCharacter { get { throw null; } }
         public override void Encode(System.IO.TextWriter output, char[] value, int startIndex, int characterCount) { }
         public override void Encode(System.IO.TextWriter output, string value, int startIndex, int characterCount) { }
         public override string Encode(string value) { throw null; }
-        public unsafe override int FindFirstCharacterToEncode(char* text, int textLength) { throw null; }
+        [System.Runtime.CompilerServices.MethodImpl(System.Runtime.CompilerServices.MethodImplOptions.AggressiveInlining)]public unsafe override int FindFirstCharacterToEncode(char* text, int textLength) { throw null; }
         public unsafe override bool TryEncodeUnicodeScalar(int unicodeScalar, char* buffer, int bufferLength, out int numberOfCharactersWritten) { throw null; }
-        public override bool WillEncode(int unicodeScalar) { throw null; }
+        [System.Runtime.CompilerServices.MethodImpl(System.Runtime.CompilerServices.MethodImplOptions.AggressiveInlining)]public override bool WillEncode(int unicodeScalar) { throw null; }
     }
     [System.AttributeUsageAttribute(System.AttributeTargets.Class, AllowMultiple=false, Inherited=false)]
     public sealed partial class OutputElementHintAttribute : System.Attribute