It's used anywhere `IHtmlContent` is rendered.

```
|                      Method |      Mean |     Error |    StdDev | Ratio | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------------------- |----------:|----------:|----------:|------:|------------:|------------:|------------:|--------------------:|
|           MvcNullHtmlEncode | 0.9833 ns | 0.0121 ns | 0.0101 ns |  1.00 |           - |           - |           - |                   - |
| InlinedBetterNullHtmlEncode | 0.2979 ns | 0.0147 ns | 0.0138 ns |  0.30 |           - |           - |           - |                   - |
```