@mkArtak 2019-02-28
Addresses https://github.com/aspnet/AspNetCore/issues/8019